Infix operator decompilation
+ and - are both binary and unary operators.  When a unary operator,
they need to be wrapped in parentheses, so that the call definition
clause is defined correctly.
